clk: sunxi: Add back fixed-clock compatible to init function
The fixed-clock handling was removed from the clk-sunxi driver when we
added the clock support for the Allwinner SoCs. We are now once again in
a position with the A31 where we don't have any clock support yet, while
we need to provide clocks to the different DT nodes.

Some unused-yet clocks in the other Allwinner SoCs also use the
fixed-clock driver (like the low frequency oscillator), and thus never
worked.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>