ARM: compressed: enable the SMP bit before decompression
On Cortex-A7 CPUs, the SMP bit needs to be enabled so that the CPU can
use its caches. However, if the bootloader doesn't set it, it's only
done after the kernel decompression, resulting in a very slow kernel
decompression.

Enable this SMP bit in compressed/head.S so that we can benefit from the
caches before decompression.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>